Refactor BoundCheckEliminationPhase to common module
Deleted duplicate BoundCheckEliminationPhase code in OpenCL, PTX, and SPIR-V modules, and created a single version in common module. This reduces redundancy and simplifies future maintenance. Updated related import statements in MidTier files accordingly.
