+++ tornado-drivers/drivers-common/src/main/java/module-info.java
@@ -1,6 +1,6 @@
 module tornado.drivers.common {
-    requires transitive jdk.internal.vm.compiler;
     requires transitive jdk.internal.vm.ci;
+    requires transitive jdk.internal.vm.compiler;
     requires transitive tornado.runtime;
 
     exports uk.ac.manchester.tornado.drivers.graal;
